Moving home screen and adding needed function to work navigation, removing launched effect for delay
