+++ app/src/main/java/com/example/soulsync/auth/HomeScreen.kt
@@ -1,9 +1,13 @@
-package com.example.soulsync
+package com.example.soulsync.auth
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun AppHome(){
+fun AppHome(
+    onNavigateToLogin: () -> Unit = {},
+    onNavigateToRegister: () -> Unit = {},
+    onNavigateToStart: () -> Unit = {}
+){
     Text(text = "Home Screen")
 }
